fix update but no images

+++ api/src/main/java/com/nocountry/pets/models/Pet.java
@@ -36,7 +36,7 @@ public class Pet {
     @Size(max = 30, message = "Colour cannot exceed 30 characters.")
     private String colour;
 
-    private Double size;
+    private Integer size;
 
     private List<String> images;
